Debugs accordion
Previous to this commit, sidenav accordion would only work on initial page load
and had behaved weirdly after going from about page back to poems. Fix by
moving accordion eventlistener attachment from within the poems button click
in nav to the entire page loading function (so that it was being called only once).
I believe it had been placing multiple listeners of the same type on the same event
and this was the cause of the issue.